+++ src/components/index.js
@@ -3,3 +3,5 @@ export {LoginForm} from './LoginForm'
 export {ContactForm} from './ContactForm'
 
 export {Todos} from './Todos'
+
+export {NotFound} from './NotFound'